Admin Arayüzü Güncellemesi

+++ src/ÇiçekDünyası.Application/DTOs/FlowerDto.cs
@@ -17,7 +17,7 @@ namespace ÇiçekDünyası.Application.DTOs
     {
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string ImageUrl { get; set; } = string.Empty;
         public int StockQuantity { get; set; }
     }
@@ -26,7 +26,7 @@ namespace ÇiçekDünyası.Application.DTOs
     {
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string ImageUrl { get; set; } = string.Empty;
         public int StockQuantity { get; set; }
         public bool IsAvailable { get; set; }
